docs(alpamayo2): complete workbench service surface
