+++ docs/cli/alpamayo2-super.md
@@ -1,0 +1,37 @@
+# `npa workbench alpamayo2-super`
+
+## Command Tree
+
+```text
+Usage: npa workbench alpamayo2-super [OPTIONS] COMMAND [ARGS]...
+
+NVIDIA Alpamayo 2 Super trajectory-inference workbench.
+
+Options
+--help  Show this message and exit.
+Commands
+infer  Run the real upstream expert trajectory inference and publish artifacts.
+terms  Print separately applicable source, model, and dataset terms.
+```
+
+## Options
+
+| Option | Description |
+| --- | --- |
+| `--help` | Show this message and exit. |
+
+## Subcommands
+
+| Command | Description |
+| --- | --- |
+| `infer` | Run the real upstream expert trajectory inference and publish artifacts. |
+| `terms` | Print separately applicable source, model, and dataset terms. |
+
+## Examples
+
+```bash
+npa workbench alpamayo2-super --help
+npa workbench alpamayo2-super infer --help
+```
+
+Regenerate this page with `bash scripts/build_docs.sh` after changing `alpamayo2-super`.
