+++ npa/src/npa/workbench/alpamayo2_super/service.py
@@ -5,65 +5,46 @@ from __future__ import annotations
 from typing import Any
 
 from fastapi import FastAPI, HTTPException
-from pydantic import BaseModel, ConfigDict
 
 from npa.workbench.alpamayo2_super.runtime import (
     ARTIFACT_SCHEMA,
-    DEFAULT_DATASET_REVISION,
-    DEFAULT_MANIFEST,
-    DEFAULT_MODEL_ID,
     DEFAULT_MODEL_REVISION,
     Alpamayo2SuperError,
     Alpamayo2SuperRequest,
     run_inference,
 )
+from npa.workbench.alpamayo2_super.schemas import InferenceBody
 
 
-class InferenceBody(BaseModel):
-    """Validated HTTP inference body."""
+def create_app() -> FastAPI:
+    """Create the deployable Alpamayo service application."""
 
-    model_config = ConfigDict(extra="forbid")
-    output_path: str
-    model_id: str = DEFAULT_MODEL_ID
-    model_revision: str = DEFAULT_MODEL_REVISION
-    dataset_revision: str = DEFAULT_DATASET_REVISION
-    manifest: str = DEFAULT_MANIFEST
-    sample_index: int = 0
-    diffusion_steps: int = 10
-    seed: int = 42
-    figure_style: str = "blog"
-    require_camera_projection: bool = True
-    run_id: str = ""
-    runtime_image: str = ""
-    dry_run: bool = False
+    service = FastAPI(title="NPA Alpamayo 2 Super", version="1.0")
 
+    @service.get("/health")
+    def health() -> dict[str, str]:
+        return {"status": "ok", "schema": ARTIFACT_SCHEMA}
 
-app = FastAPI(title="NPA Alpamayo 2 Super", version="1.0")
+    @service.get("/system-info")
+    def system_info() -> dict[str, Any]:
+        return {"model_revision": DEFAULT_MODEL_REVISION, "weights_baked": False}
 
+    @service.post("/run")
+    def run(body: InferenceBody) -> dict[str, Any]:
+        try:
+            return run_inference(Alpamayo2SuperRequest(**body.model_dump()))
+        except Alpamayo2SuperError as exc:
+            raise HTTPException(status_code=422, detail=str(exc)) from exc
 
-@app.get("/health")
-def health() -> dict[str, str]:
-    return {"status": "ok", "schema": ARTIFACT_SCHEMA}
+    @service.get("/status")
+    def status() -> dict[str, str]:
+        return {"status": "ready"}
 
+    @service.get("/list")
+    def list_capabilities() -> list[dict[str, Any]]:
+        return [{"name": "trajectory-inference", "schema": ARTIFACT_SCHEMA}]
 
-@app.get("/system-info")
-def system_info() -> dict[str, Any]:
-    return {"model_revision": DEFAULT_MODEL_REVISION, "weights_baked": False}
-
-
-@app.post("/run")
-def run(body: InferenceBody) -> dict[str, Any]:
-    try:
-        return run_inference(Alpamayo2SuperRequest(**body.model_dump()))
-    except Alpamayo2SuperError as exc:
-        raise HTTPException(status_code=422, detail=str(exc)) from exc
+    return service
 
 
-@app.get("/status")
-def status() -> dict[str, str]:
-    return {"status": "ready"}
-
-
-@app.get("/list")
-def list_capabilities() -> list[dict[str, Any]]:
-    return [{"name": "trajectory-inference", "schema": ARTIFACT_SCHEMA}]
+app = create_app()
